fix : StringAddCalculator Test Code 수정(@Nested 활용)
